+++ test/globals.js
@@ -112,6 +112,9 @@ const fundingEligibility = {
     itau: {
         eligible: false
     },
+    satispay: {
+        eligible: false
+    },
     paidy: {
         eligible: false
     }