+++ test/integration/globals.js
@@ -112,6 +112,9 @@ window.__TEST_FUNDING_ELIGIBILITY__ = {
     applepay: {
         eligible: true
     },
+    satispay: {
+        eligible: false
+    },
     paidy: {
         eligible: false
     }