+++ test/integration/tests/utils/index.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { COUNTRY, CURRENCY, FUNDING } from '@paypal/sdk-constants/src';
+import { CURRENCY, FUNDING } from '@paypal/sdk-constants/src';
 
 import { BUTTON_LAYOUT } from '../../../../src/constants';
 import { isInlineXOEligible } from '../../../../src/zoid/buttons/util';
@@ -26,6 +26,7 @@ describe('isInlineXOEligible', () => {
         props = {
             commit:             true,
             currency:           CURRENCY.USD,
+            custom:             { label: 'Checkout', css: {} },
             disableFunding:     [ FUNDING.CREDIT ],
             fundingEligibility: {
                 [FUNDING.CARD]: {
@@ -39,7 +40,6 @@ describe('isInlineXOEligible', () => {
                 }
             },
             layout: BUTTON_LAYOUT.VERTICAL,
-            locale: { country: COUNTRY.US, lang: 'en' },
             onComplete: () => true,
             vault:  false
         };
@@ -50,15 +50,7 @@ describe('isInlineXOEligible', () => {
             throw new Error(`Expected ${ JSON.stringify(props) } to be eligible.`);
         }
     });
-   
-    it('should be ineligible if locale.country is not US', () => {
-        props.locale.country = COUNTRY.DE;
 
-        if (isInlineXOEligible({ props })) {
-            throw new Error(`Expected buyerCountry, ${ COUNTRY.DE } to be ineligible.`);
-        }
-    });
-   
     it('should be ineligible if commit is false', () => {
         props.commit = false;
 
@@ -99,6 +91,14 @@ describe('isInlineXOEligible', () => {
 
         if (isInlineXOEligible({ props })) {
             throw new Error(`Expected layout=horizontal to be ineligible.`);
+        }
+    });
+   
+    it('should be ineligible if custom is not set', () => {
+        props.custom = undefined;
+
+        if (isInlineXOEligible({ props })) {
+            throw new Error(`Expected custom style to be set to be eligible.`);
         }
     });